requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy<2.0
sentence-transformers
streamlit
PyPDF2
spacy
faiss-cpu
keybert
matplotlib